:truck: refactor: move home page section widgets into own folder
