+++ client/flutter/lib/components/home_page_sections/home_page_information_card.dart
